refactor: move rooms repository+mappers into core-rooms, add observeRooms Flow
- Add observeAllRoomsWithStudents(): Flow<List<RoomWithStudents>> to RoomDao
- Move RoomsRepositoryImpl, RoomMapper, StudentMapper from feature-rooms:impl to core-rooms
- Implement RoomsRepository.observeRooms() via Room DAO Flow query
- Create coreRoomsModule Koin module in core-rooms:di
- Delete originals from feature-rooms:impl (dangling refs fixed in T4)
