+++ shared/core-rooms/src/commonMain/kotlin/dev/nonoxy/residetrack/core/rooms/data/mappers/StudentMapper.kt
@@ -1,9 +1,9 @@
-package dev.nonoxy.residetrack.feature.rooms.impl.data.mappers
+package dev.nonoxy.residetrack.core.rooms.data.mappers
 
 import dev.nonoxy.residetrack.common.utils.mapper.Mapper
 import dev.nonoxy.residetrack.common.utils.toLocalDate
 import dev.nonoxy.residetrack.core.database.entities.StudentEntity
-import dev.nonoxy.residetrack.feature.rooms.api.models.Student
+import dev.nonoxy.residetrack.core.rooms.models.Student
 import kotlin.time.Clock
 import kotlinx.datetime.TimeZone
 import kotlinx.datetime.atStartOfDayIn
